@@ -1,19 +1,52 @@
-var webpack = require('webpack');
+/* global __dirname, require, module*/
 
-module.exports = {
-	entry		: './src/iAjax.js',
-	output		: {
-		path			: __dirname+'/dist',
-		filename		: 'iAjax.js',
-		library			: 'iAjax',
-		libraryTarget	: 'var'		// amd
-	},
-	module		: {
-		loaders		: [
-			{ text: /\.js$/, exclude: /node_modules/, loader: 'babel-loader' }
-		]
-	},
-	plugins		: [
-		new webpack.optimize.UglifyJsPlugin({ sourceMap: true })
-	]
+const webpack = require('webpack');
+const path = require('path');
+const env = require('yargs').argv.env; // use --env with webpack 2
+const pkg = require('./package.json');
+
+let libraryName = pkg.name;
+
+let outputFile, mode;
+
+if (env === 'build') {
+  mode = 'production';
+  outputFile = libraryName + '.min.js';
+} else {
+  mode = 'development';
+  outputFile = libraryName + '.js';
+}
+
+const config = {
+  mode: mode,
+  entry: __dirname + '/src/iAjax.js',
+  // devtool: 'inline-source-map',
+  output: {
+    path: __dirname + '/lib',
+    filename: outputFile,
+    library: libraryName,
+    libraryTarget: 'umd'
+    // umdNamedDefine: true,
+    // globalObject: "typeof self !== 'undefined' ? self : this"
+  },
+  module: {
+    rules: [
+      {
+        test: /(\.jsx|\.js)$/,
+        loader: 'babel-loader',
+        exclude: /(node_modules|bower_components)/
+      }/*,
+      {
+        test: /(\.jsx|\.js)$/,
+        loader: 'eslint-loader',
+        exclude: /node_modules/
+      }*/
+    ]
+  },
+  resolve: {
+    modules: [path.resolve('./node_modules'), path.resolve('./src')],
+    extensions: ['.json', '.js']
+  }
 };
+
+module.exports = config;
